chore(utils): fix circular dependency (#12110)
* chore: fix circular dependency

* chore: fix

* chore: fix test
