+++ packages/vant/src/utils/basic.ts
@@ -1,4 +1,3 @@
-import { isObject } from './validate';
 import type { ComponentPublicInstance } from 'vue';
 
 export function noop() {}
@@ -11,6 +10,38 @@ export type Numeric = number | string;
 
 // eslint-disable-next-line
 export type ComponentInstance = ComponentPublicInstance<{}, any>;
+
+export const isObject = (val: unknown): val is Record<any, any> =>
+  val !== null && typeof val === 'object';
+
+export const isDef = <T>(val: T): val is NonNullable<T> =>
+  val !== undefined && val !== null;
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export const isFunction = (val: unknown): val is Function =>
+  typeof val === 'function';
+
+export const isPromise = <T = any>(val: unknown): val is Promise<T> =>
+  isObject(val) && isFunction(val.then) && isFunction(val.catch);
+
+export const isDate = (val: unknown): val is Date =>
+  Object.prototype.toString.call(val) === '[object Date]' &&
+  !Number.isNaN((val as Date).getTime());
+
+export function isMobile(value: string): boolean {
+  value = value.replace(/[^-|\d]/g, '');
+  return (
+    /^((\+86)|(86))?(1)\d{10}$/.test(value) || /^0[0-9-]{10,13}$/.test(value)
+  );
+}
+
+export const isNumeric = (val: Numeric): val is string =>
+  typeof val === 'number' || /^\d+(\.\d+)?$/.test(val);
+
+export const isIOS = (): boolean =>
+  inBrowser
+    ? /ios|iphone|ipad|ipod/.test(navigator.userAgent.toLowerCase())
+    : false;
 
 export function get(object: any, path: string): any {
   const keys = path.split('.');
@@ -32,14 +63,17 @@ export type RequiredParams<T> = T extends (...args: infer P) => infer R
 export function pick<T, U extends keyof T>(
   obj: T,
   keys: ReadonlyArray<U>,
-  ignoreUndefined?: boolean
+  ignoreUndefined?: boolean,
 ) {
-  return keys.reduce((ret, key) => {
-    if (!ignoreUndefined || obj[key] !== undefined) {
-      ret[key] = obj[key];
-    }
-    return ret;
-  }, {} as Writeable<Pick<T, U>>);
+  return keys.reduce(
+    (ret, key) => {
+      if (!ignoreUndefined || obj[key] !== undefined) {
+        ret[key] = obj[key];
+      }
+      return ret;
+    },
+    {} as Writeable<Pick<T, U>>,
+  );
 }
 
 export const isSameValue = (newValue: unknown, oldValue: unknown) =>
