+++ packages/vant/src/utils/create.ts
@@ -1,6 +1,5 @@
-import { get } from './basic';
+import { get, isFunction } from './basic';
 import { camelize } from './format';
-import { isFunction } from './validate';
 import locale from '../locale';
 
 export function createTranslate(name: string) {
@@ -31,13 +30,13 @@ function genBem(name: string, mods?: Mods): string {
   if (Array.isArray(mods)) {
     return (mods as Mod[]).reduce<string>(
       (ret, item) => ret + genBem(name, item),
-      ''
+      '',
     );
   }
 
   return Object.keys(mods).reduce(
     (ret, key) => ret + (mods[key] ? genBem(name, key) : ''),
-    ''
+    '',
   );
 }
